requirements.txt
aiocache==0.11.1
aiocqhttp==1.2.5
aiofiles==0.5.0
APScheduler==3.6.3
beautifulsoup4==4.9.0
blinker==1.4
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.4.28
httpx==0.12.1
Hypercorn==0.9.5
hyperframe==5.2.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
nonebot==1.6.0
Pillow==7.1.2
priority==1.3.0
pytz==2020.1
Quart==0.11.5
requests==2.23.0
rfc3986==1.4.0
six==1.14.0
sniffio==1.1.0
soupsieve==2.0
toml==0.10.0
typing-extensions==3.7.4.2
tzlocal==2.0.0
urllib3==1.25.9
Werkzeug==1.0.1
wsproto==0.15.0
